refactor: move DATA_THEMES to config file
