+++ app/config.ts
@@ -1,3 +1,42 @@
 export const DATASET_CATALOG_PATH = 'data-catalog';
 export const STORY_HUB_PATH = 'stories';
 export const EXPLORATION_PATH = 'exploration';
+
+export const DATA_THEMES = [
+  {
+    id: 'agriculture',
+    title: 'Agriculture',
+  },
+  {
+    id: 'air-quality',
+    title: 'Air Quality',
+  },
+  {
+    id: 'biodiversity',
+    title: 'Biodiversity',
+  },
+  {
+    id: 'disasters',
+    title: 'Disasters',
+  },
+  {
+    id: 'energy',
+    title: 'Energy',
+  },
+  {
+    id: 'greenhouse-gases',
+    title: 'Greenhouse Gases',
+  },
+  {
+    id: 'sea-level-rise',
+    title: 'Sea Level Rise',
+  },
+  {
+    id: 'wildfires',
+    title: 'Wildfires',
+  },
+  {
+    id: 'water-resources',
+    title: 'Water Resources',
+  },
+] as const;
